fix flash firmware
fashing a new firmware was not possible.
the device requested always the first 4096 bytes of the new firmware.

the http header names are case insensitive. so the most rust
web-frameworks uses the 'http' crate. this crate serves all
header names lower-case.

so i implemented a minimal handmade http server.
if i send lower-case http headers, the sonoff device
request only the first 4096 bytes. when is send
the names with the first letter in upper case - it works!